+++ app/Http/Controllers/MarketController.php
@@ -97,7 +97,7 @@ class MarketController extends Controller
                     $imagePath = public_path($market->market_image); // Get the full image path
 
                     // Delete the image file if it exists
-                    if (file_exists($imagePath)) {
+                    if (!empty($market->market_image) && file_exists($imagePath) && is_file($imagePath)) {
                         unlink($imagePath); // Delete the image from the file system
                     }
